update package name for lint error
