+++ security-command-center/snippets/src/main/java/v_two/muteconfig/CreateMuteRule.java
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package v2.muteconfig;
+package v_two.muteconfig;
 
 // [START securitycenter_create_mute_config_v2]
