+++ security-command-center/snippets/src/main/java/v_two/muteconfig/DeleteMuteRule.java
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package v2.muteconfig;
+package v_two.muteconfig;
 
 // [START securitycenter_delete_mute_config_v2]
