+++ security-command-center/snippets/src/main/java/v_two/muteconfig/SetUnmuteFinding.java
@@ -14,9 +14,9 @@
  * limitations under the License.
  */
 
-package v2.muteconfig;
+package v_two.muteconfig;
 
-// [START securitycenter_set_mute_v2]
+// [START securitycenter_set_unmute_v2]
 
 import com.google.cloud.securitycenter.v2.Finding;
 import com.google.cloud.securitycenter.v2.Finding.Mute;
@@ -24,7 +24,7 @@ import com.google.cloud.securitycenter.v2.SecurityCenterClient;
 import com.google.cloud.securitycenter.v2.SetMuteRequest;
 import java.io.IOException;
 
-public class SetMuteFinding {
+public class SetUnmuteFinding {
 
   public static void main(String[] args) throws IOException {
     // TODO: Replace the variables within {}
@@ -37,22 +37,21 @@ public class SetMuteFinding {
     //
     String findingPath = "{path-to-the-finding}";
 
-    setMute(findingPath);
+    setUnmute(findingPath);
   }
 
-  // Mute an individual finding.
-  // If a finding is already muted, muting it again has no effect.
+  // Unmute an individual finding.
+  // Unmuting a finding that isn't muted has no effect.
   // Various mute states are: MUTE_UNSPECIFIED/MUTE/UNMUTE.
-  public static void setMute(String findingPath) throws IOException {
+  public static void setUnmute(String findingPath) throws IOException {
     // Initialize client that will be used to send requests. This client only needs to be created
     // once, and can be reused for multiple requests.
     try (SecurityCenterClient client = SecurityCenterClient.create()) {
 
       SetMuteRequest setMuteRequest =
           SetMuteRequest.newBuilder()
-              // Relative path for the finding.
               .setName(findingPath)
-              .setMute(Mute.MUTED)
+              .setMute(Mute.UNMUTED)
               .build();
 
       Finding finding = client.setMute(setMuteRequest);
@@ -61,4 +60,4 @@ public class SetMuteFinding {
     }
   }
 }
-// [END securitycenter_set_mute_v2]
+// [END securitycenter_set_unmute_v2]
