+++ security-command-center/snippets/src/test/java/v_two/MuteFindingIT.java
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package v2;
+package v_two;
 
 import static com.google.common.truth.Truth.assertThat;
 import static com.google.common.truth.Truth.assertWithMessage;
@@ -44,14 +44,14 @@ import org.junit.Rule;
 import org.junit.Test;
 import org.junit.runner.RunWith;
 import org.junit.runners.JUnit4;
-import v2.muteconfig.BulkMuteFindings;
-import v2.muteconfig.CreateMuteRule;
-import v2.muteconfig.DeleteMuteRule;
-import v2.muteconfig.GetMuteRule;
-import v2.muteconfig.ListMuteRules;
-import v2.muteconfig.SetMuteFinding;
-import v2.muteconfig.SetUnmuteFinding;
-import v2.muteconfig.UpdateMuteRule;
+import v_two.muteconfig.BulkMuteFindings;
+import v_two.muteconfig.CreateMuteRule;
+import v_two.muteconfig.DeleteMuteRule;
+import v_two.muteconfig.GetMuteRule;
+import v_two.muteconfig.ListMuteRules;
+import v_two.muteconfig.SetMuteFinding;
+import v_two.muteconfig.SetUnmuteFinding;
+import v_two.muteconfig.UpdateMuteRule;
 
 // Test v2 Mute config samples.
 @RunWith(JUnit4.class)
