+++ security-command-center/snippets/src/test/java/v_two/Util.java
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package v2;
+package v_two;
 
 import com.google.cloud.securitycenter.v2.CreateFindingRequest;
 import com.google.cloud.securitycenter.v2.CreateSourceRequest;
